HDDS-14648. Add a prefix to SCM codec classes (#9782)
