+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/ha/io/ScmBigIntegerCodec.java
@@ -17,26 +17,23 @@
 
 package org.apache.hadoop.hdds.scm.ha.io;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import java.math.BigInteger;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
-import org.junit.jupiter.api.Test;
+import org.apache.ratis.thirdparty.com.google.protobuf.UnsafeByteOperations;
 
 /**
- * Class to test BigIntegerCodec serialize and deserialize.
+ * Codec for type BigInteger.
  */
-public class TestBigIntegerCodec {
+public class ScmBigIntegerCodec implements ScmCodec {
 
-  @Test
-  public void testCodec() {
-    BigIntegerCodec bigIntegerCodec = new BigIntegerCodec();
+  @Override
+  public ByteString serialize(Object object) {
+    // BigInteger returns a new byte[].
+    return UnsafeByteOperations.unsafeWrap(((BigInteger) object).toByteArray());
+  }
 
-    BigInteger bigInteger = BigInteger.valueOf(100);
-    ByteString byteString = bigIntegerCodec.serialize(bigInteger);
-
-    BigInteger actual =
-        (BigInteger) bigIntegerCodec.deserialize(BigInteger.class, byteString);
-    assertEquals(bigInteger, actual);
+  @Override
+  public Object deserialize(Class< ? > type, ByteString value) {
+    return new BigInteger(value.toByteArray());
   }
 }
