+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/ha/io/ScmManagedSecretKeyCodec.java
@@ -17,36 +17,34 @@
 
 package org.apache.hadoop.hdds.scm.ha.io;
 
-import com.google.protobuf.Message;
-import java.lang.reflect.InvocationTargetException;
-import org.apache.hadoop.hdds.scm.ha.ReflectionUtil;
+import org.apache.hadoop.hdds.protocol.proto.SCMSecretKeyProtocolProtos;
+import org.apache.hadoop.hdds.security.symmetric.ManagedSecretKey;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
 import org.apache.ratis.thirdparty.com.google.protobuf.InvalidProtocolBufferException;
 import org.apache.ratis.thirdparty.com.google.protobuf.UnsafeByteOperations;
 
 /**
- * {@link Codec} implementation for non-shaded
- * {@link com.google.protobuf.Message} objects.
+ * A codec for {@link ManagedSecretKey} objects.
  */
-public class GeneratedMessageCodec implements Codec {
-
+public class ScmManagedSecretKeyCodec implements ScmCodec {
   @Override
   public ByteString serialize(Object object)
       throws InvalidProtocolBufferException {
+    ManagedSecretKey secretKey = (ManagedSecretKey) object;
     return UnsafeByteOperations.unsafeWrap(
-        ((Message) object).toByteString().asReadOnlyByteBuffer());
+        secretKey.toProtobuf().toByteString().asReadOnlyByteBuffer());
   }
 
   @Override
   public Object deserialize(Class<?> type, ByteString value)
       throws InvalidProtocolBufferException {
     try {
-      return ReflectionUtil.getMethod(type, "parseFrom", byte[].class)
-          .invoke(null, (Object) value.toByteArray());
-    } catch (NoSuchMethodException | IllegalAccessException
-             | InvocationTargetException ex) {
-      ex.printStackTrace();
-      throw new InvalidProtocolBufferException("Message cannot be decoded: " + ex.getMessage());
+      SCMSecretKeyProtocolProtos.ManagedSecretKey message =
+          SCMSecretKeyProtocolProtos.ManagedSecretKey.parseFrom(
+              value.asReadOnlyByteBuffer());
+      return ManagedSecretKey.fromProtobuf(message);
+    } catch (com.google.protobuf.InvalidProtocolBufferException e) {
+      throw new InvalidProtocolBufferException("Failed to deserialize value for " + type, e);
     }
   }
 }
