+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/ha/io/ScmNonShadedByteStringCodec.java
@@ -17,42 +17,27 @@
 
 package org.apache.hadoop.hdds.scm.ha.io;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-import java.security.cert.X509Certificate;
-import org.apache.hadoop.hdds.security.x509.certificate.utils.CertificateCodec;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
 import org.apache.ratis.thirdparty.com.google.protobuf.InvalidProtocolBufferException;
 import org.apache.ratis.thirdparty.com.google.protobuf.UnsafeByteOperations;
 
 /**
- * Codec for type X509Certificate.
+ * {@link ScmCodec} implementation for non-shaded
+ * {@link com.google.protobuf.ByteString} objects.
  */
-public class X509CertificateCodec implements Codec {
+public class ScmNonShadedByteStringCodec implements ScmCodec {
+
   @Override
   public ByteString serialize(Object object)
       throws InvalidProtocolBufferException {
-    try {
-      String certString =
-          CertificateCodec.getPEMEncodedString((X509Certificate) object);
-      // getBytes returns a new array
-      return UnsafeByteOperations.unsafeWrap(certString.getBytes(UTF_8));
-    } catch (Exception ex) {
-      throw new InvalidProtocolBufferException(
-          "X509Certificate cannot be decoded: " + ex.getMessage());
-    }
+    return UnsafeByteOperations.unsafeWrap(
+        ((com.google.protobuf.ByteString) object).asReadOnlyByteBuffer());
   }
 
   @Override
-  public Object deserialize(Class< ? > type, ByteString value)
+  public Object deserialize(Class<?> type, ByteString value)
       throws InvalidProtocolBufferException {
-    try {
-      String pemEncodedCert = new String(value.toByteArray(), UTF_8);
-      return CertificateCodec.getX509Certificate(pemEncodedCert);
-    } catch (Exception ex) {
-      throw new InvalidProtocolBufferException(
-          "X509Certificate cannot be decoded: " + ex.getMessage());
-    }
+    return com.google.protobuf.UnsafeByteOperations.
+        unsafeWrap(value.asReadOnlyByteBuffer());
   }
 }
-
