+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/ha/io/ScmNonShadedGeneratedMessageCodec.java
@@ -17,27 +17,36 @@
 
 package org.apache.hadoop.hdds.scm.ha.io;
 
+import com.google.protobuf.Message;
+import java.lang.reflect.InvocationTargetException;
+import org.apache.hadoop.hdds.scm.ha.ReflectionUtil;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
 import org.apache.ratis.thirdparty.com.google.protobuf.InvalidProtocolBufferException;
 import org.apache.ratis.thirdparty.com.google.protobuf.UnsafeByteOperations;
 
 /**
- * {@link Codec} implementation for non-shaded
- * {@link com.google.protobuf.ByteString} objects.
+ * {@link ScmCodec} implementation for non-shaded
+ * {@link com.google.protobuf.Message} objects.
  */
-public class ByteStringCodec implements Codec {
+public class ScmNonShadedGeneratedMessageCodec implements ScmCodec {
 
   @Override
   public ByteString serialize(Object object)
       throws InvalidProtocolBufferException {
     return UnsafeByteOperations.unsafeWrap(
-        ((com.google.protobuf.ByteString) object).asReadOnlyByteBuffer());
+        ((Message) object).toByteString().asReadOnlyByteBuffer());
   }
 
   @Override
   public Object deserialize(Class<?> type, ByteString value)
       throws InvalidProtocolBufferException {
-    return com.google.protobuf.UnsafeByteOperations.
-        unsafeWrap(value.asReadOnlyByteBuffer());
+    try {
+      return ReflectionUtil.getMethod(type, "parseFrom", byte[].class)
+          .invoke(null, (Object) value.toByteArray());
+    } catch (NoSuchMethodException | IllegalAccessException
+             | InvocationTargetException ex) {
+      ex.printStackTrace();
+      throw new InvalidProtocolBufferException("Message cannot be decoded: " + ex.getMessage());
+    }
   }
 }
