+++ hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/ha/io/TestScmBigIntegerCodec.java
@@ -17,27 +17,26 @@
 
 package org.apache.hadoop.hdds.scm.ha.io;
 
-import com.google.common.primitives.Longs;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.math.BigInteger;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
-import org.apache.ratis.thirdparty.com.google.protobuf.InvalidProtocolBufferException;
-import org.apache.ratis.thirdparty.com.google.protobuf.UnsafeByteOperations;
+import org.junit.jupiter.api.Test;
 
 /**
- * {@link Codec} for {@code Long} objects.
+ * Class to test BigIntegerCodec serialize and deserialize.
  */
-public class LongCodec implements Codec {
+public class TestScmBigIntegerCodec {
 
-  @Override
-  public ByteString serialize(Object object)
-      throws InvalidProtocolBufferException {
-    // toByteArray returns a new array
-    return UnsafeByteOperations.unsafeWrap(Longs.toByteArray((Long) object));
+  @Test
+  public void testCodec() {
+    ScmBigIntegerCodec scmBigIntegerCodec = new ScmBigIntegerCodec();
+
+    BigInteger bigInteger = BigInteger.valueOf(100);
+    ByteString byteString = scmBigIntegerCodec.serialize(bigInteger);
+
+    BigInteger actual =
+        (BigInteger) scmBigIntegerCodec.deserialize(BigInteger.class, byteString);
+    assertEquals(bigInteger, actual);
   }
-
-  @Override
-  public Object deserialize(Class<?> type, ByteString value)
-      throws InvalidProtocolBufferException {
-    return Longs.fromByteArray(value.toByteArray());
-  }
-
 }
